Convert to a client app approach
This will utilize react and the client will live external to the
theme.  It compiles via webpack before being placed bundled in
the theme directory ultimately placed into the public directory
for consumption.

Perhaps hexo-react-rendered will be able to render out the initial
bit and allow use of the hexo variables.